Day 3: Rucksack Reorganization
Fun day my initial approach was to count the number of each item
in the rucksack and then manually find the intersection.

When part two opened up I noticed the item count was not needed
so I used sets instead. I also realized it's easier to use set
operations to find the intersection rather than looping through
the items. Turned out to be a pretty clean and easy solution.