2022 Day 8: Treetop Tree House
I had some problem with finding a good way of finding which trees were
visible and which were not in the beginning.

Then another challenge was to refactor the code to be able to reuse
similar parts. There's still a lot of similarities between part 1
and part 2, but sharing even more code could make it harder to
understand what's going on.