2022 Day 11: Monkey in the Middle
First part was fairly simple, but I got worried for a while when the
example input worked and all parts was correct but the real input
didn't work. Then I thought about Int overflow and changed Int to Long
and the answer was accepted.

Part two felt tricky at first, but after thinking a bit I realized
I could use a modulo to keep the numbers from growing too much.
Then it was a fairly simple modification to get it working.