2022 Day 13: Distress Signal
Easy to know what to do, but I'm not good at all with
using a stack to parse this kind of data. Tried different
approaches for almost an hour before I looked at how I
did exactly the same parsing in 2015 day 12. After that
it was simple.

Part two was fairly easy, but I had to Google a bit to
find the name and usage of 'sortedWith'.