Fix BFS search
Do this by including the start position in the history of
visited nodes and make it point to itself. If it's not included
the Result.getPath() function will get stuck in an infinite
loop as the first neighbour of the start position will lead back
to the start position again.

Also add a few simple tests that catches this problem with BFS,
as well as improve documentation on how to use these functions.