2022 Day 14: Regolith Reservoir
Simulating the first part was fun and fairly easy. But for some
reason it takes around 10 seconds to find the answer. Tried the
same solution for part two, but that took about 1h10 minutes to
finish.

So I instead caclulated the amount of sand by blocking out all
positions where sand could not end up and subtracted that from
the maximum amount of sand that would have landed if there
would have not been any rocks. That finds the solution in
100ms.

I need to get back to this later and find out what's taking
time and find a faster solution for the first part. Other
people on Reddit doesn't seem to have problems with the
solution being extremely slow.