2022 Day 15: Beacon Exclusion Zone
It was fairly straight forward to understand what to do, but I still struggled
a lot with this one.

First approach was to look at the target row and try to create a list of
IntRanges for all the # in the illustration in the puzzle. This required
to first generate an initial range and then punch a lot of holes in it
where there were beacons and sensors. This was fairly complicated and buggy.

When I came to part 2 I had to take the range with holes and plug the holes
which also was a bit complicated.

After I solved it, I realized that it's better to just have a single range
instead of one with holes in it. That made everything much simpler, faster
and I could delete almost half of my code.

Another thing I had a lot of problems with was to write the algorithms
for doing unions on IntRanges and even more difficult, unions of a whole
list of IntRanges. But now I have those as utility functions so I shouldn't
have to struggle with that again in the future.