2022 Day 17: Pyroclastic Flow
First part was fairly easy. Had a few small bugs due to missunderstanding
the coordinates of the blocks but it was easy to spot and fix.

For the second it was obvious that there were loops to be found but
it took a little while to figure out how to detect them. Then I
spent 1-2 hours trying to figure out how to calculate the height
right. A lot of off-by-one errors back and forth before I found the
right approach.

Not a completely generic solution, here I'm assuming that the loop
will start before rock 200, be shorter than 1800 and that there will
never be sequences of more than 10 similar height differences that's
not part of the loop. Works fine for my input, but could break down
with other inputs.