2022 Day 20: Grove Positioning System
I immediately suspected that part two would require a lot of mixing
and that some kind of linked list approach would be required.

But just in case I tried to do a simple normal list approach where
I just removed and inserted items by index. Worked fine with the
example input, but didn't work with the real input since the real
input was much more complicated, it both contained several identical
numbers and and movements that were close to 2 times larger than
the size of the circle.

So I had to implement a linked list approach which wasn't too
difficult. The most difficult / error prone part was to calculate the
minimum required steps to move a much larger amount of steps. But with
several unit tests it was possible to come up with a correct implementation.

When I finished part one my code was in fairly good shape for part two,
it just needed a bit of refactoring to be able to mix several time and
keep the order from before the first mixing. But that was fairly straight
forward.

All in all, it was a bit tricky to figure out a good implementation but
it was a fun challenge.