2022 Day 16: Proboscidea Volcanium
This was really difficult for me. When reading it it felt like it
should be easy. Just do a graph of the tunnels and valves and run
djikstra or a* on it.

But then it turned out that creating the graph wasn't as easy as I
thought. After some time I solved part 1 and part 2 felt like it would
only require some changes to the graph. But doing that made it run out
of heap space very quickly, the amount of different steps were just too
big.

After some thinking I decided to remove all nodes that you can only
pass through. Then there was a lot of debugging and struggling before
I finally got a working implementation that passed the test case. I had
a lot of small issues and debugging where search algorithms go wrong
is not easy.

Running it on the real input still ran out of heap space so that didn't
go well. Had to resort to a heuristic where I only ignore turning valves
with a low pressure rate. This cut down the search space enough to be
able to solve part 2 in a couple of seconds.

This solution is pretty messy, so I'm going to try to improve it in
a future commit based on ideas I got when reading some other solutions.