2022 Day 19: Not Enough Minerals
This was another complicated day. The basic idea is fairly easy: do
BFS, djikstra or a* to find the best sequence. Not particularly
difficult to make a graph over the different steps either, but then
when you try to run it the number of possible states is just too large.

So the big challenge is how to find ways of not adding useless states
and how to do a fairly reasonable estimate of the final state.

After a lot of thinking I could get the first star, but for the second
I needed a few small hints on further optimizations. Even with that
I only managed to solve it if I put in some very input dependent
restrictions.

Once I had the solution I spent quite some time to rewrite the solution
to a DFS approach instead that only evaluates a branch if it has a
chance to beat the current best. This increased speed from several seconds
to around 200ms.

The final challenge was then to refactor the code to look a bit nicer and
avoid having a state with named parameters for each resource and bot that
you have to write out everywhere. Still a bit messy, but at least better
than what I had before.