Solve 2022 day 16 in a cleaner and faster way
Reduce the graph so that it only contains the nodes of interest where
each node is walking to a valve and also opening it. Then calculate
the distances between all nodes.

Took qiute a lot of time to iron out all bugs and mistakes. Initially
this solution ran out of heap space on part 2, but after I improved
the heuristic a bit it finished in around 400ms with part 1 taking 20ms.