2023 Day 23: Unstable Diffusion
Fun variant of game of life. Felt like a good idea to just store the
positions of the elves and no empty spaces from the start. When I got
to part 2 it turned out that was the correct decision, so part 2
was pretty eays.

Part two takes around 1.3-1.5 seconds for me, which is a bit more than
I hoped for. Found some small optimizations that cut off a couple of
hundred ms from the runtime.

Can't really get it faster than this without rewriting it completely
and avoid using sets, which would make the code much more complicated.

This will have to be good enough.