Use a simpler approach for 2022 day 24
Instead of generating all blizzard states I just look in four direction on
the initial state to see if there is a blizzard that will end up on the
current position at a given time. Takes more or less the same time as the
previous approach, so I guess the lookups are taking most time rather than
generating blizzard states.

But I'm keeping this solution as it's smaller and easier to understand.