2023 Day 6: Wait For It
Had most problems with the parsing. Tried using regexp and ran in to
the classic where I tried to use regex to solve a problem just to find
my self with two problems.

Scrapped that idea and tried to split on a regex. But the same thing
again, that caused more problems. But that was fixed with a bit of
googling.

Solving the problem was easy after the parsing was done. I thought part
2 would require some clever approach due to the large numbers but I did
a naive implementation and was surprised to see that it finished in a few
tenths of milliseconds.