2023 Day 11: Cosmic Expansion
Felt like it would be a lot to do, but most of the work was to
expand the galaxy. Creating all pairs was trivial and finding
the minimum distances was just manhattan distances. The description
made it sound like some kind of BFS variant would be needed.

Having a sparse map with only the points of interest included and
not empty space made part 2 trivial. No problems with increasing the
amount of empty space.