2023 Day 12: Hot Springs (part 1)
This was difficult. Got part 1 working with a BFS approach and a
lot of debugging. Ran out of memory within a few seconds on part 2.

So rewrote it to a DFS instead. But that runs for 10+ hours without
finding a solution, so something more clever have to be done for
part 2.