2023 Day 14: Parabolic Reflector Dish
Part one was fairly straightforward and fun. For part two it was
obvious that you had to find loops. Did a quick copy paste implementation
where each tilt direction is a lot of code duplication.

As feared, using AMap and creating new objects for every rotation makes
things slow and part two takes about 1.5 minutes to solve.

Need to come back to this later and change it to a Grid instead and
try to work with only 2 copies to reduce the amount of allocations needed.
And after that also fix the tilt functions to not have 4 almost identical
functions.