2023 Day 18: Lavaduct Lagoon
Part 1 was easy, part 2 felt like it should be easy as well.
"Just do a scan line algorithm" was my thought. In principle
that was easy but it was the first time I implemented something
like that and it took forever to get it right.

Need to come back later and clean up the code.