2023 Day 23: A Long Walk
Part one was fairly easy and fun thanks to some valuable learnings
last year about reducing a map to only the necessary parts.

Part two was really difficult. Tried a BFS approach first but it
ran out of memory a few steps from finishing. Tweaked it a few times
and got further, but not quite far enough.

Changed to a DFS approach instead which also ran out of memory. With
some optimizations I managed to get it to barely finish only to find
that I had a too aggresive optimization so that I missed some cases.

Rewrote my solution to look at edges instead of nodes and managed to
get a solution that worked and with a couple of rewrites and further
optimizations I got it to run in 1.5 seconds.

Part of the code was pretty heavy and hard to follow so tried
rewriting it to use nodes instead of edges. Made many parts much
simpler and quicker, however the heuristic for remaining path were
more difficult. Same heuristic as before took 10 seconds  instead of
a few hundred milliseconds.

Had to settle with a worse heuristic that only took a few hundred ms
but caused the algorithm to visit almost twice the amount of nodes.
Overall runtime ended up close to 1.5 seconds here as well, but the
code was simpler so I stuck with this solution.

Key optimizations / heuristics:
 - If moving to a new node means one of the nodes that's not visited
   becomes a dead end, remove that dead end.
 - Estimate remaining cost by taking the best available edge leading
   away from each node.