2023 Day 24: Never Tell Me The Odds
Part 1 was fairly easy where I could find an algorithm
on Wikipedia for finding the intersection of two lines
in 2D.

Part 2 was much more difficult. Spent a lot of time trying
to figure out something (non-math:y) on my own but failed
so had to resort to Reddit for some help. Saw Gaussian
elimination being mentioned so tried to implement that which
was a challenge.

Went with a Gauss-Jordan variant that turned out to have too
many rounding errors. Asked on reddit and got a suggestion for
a Gaussian elimination variant that worked solely on integers.

With that in place I managed to finally solve part 2 and if the need
to solve linear equations shows up again in the future I can hopefully
remember about this and re-use the Gaussian elimination method i wrote.