2023 Day 21: Step Counter Part 2
This was a real challenge. Managed to split up the huge shape into
many smaller parts and count the different parts separately. Were
a lot of cases to consider and a lot of thinking to figure out how
it works. I had some annoying overflow issues that took a while to
track down in the end.

People who were better at math had much better solutions than me, but
I'm proude that I could come up with this completely on my own and
that it worked.

Doing a lot of visualizations helped figure out how to solve it, so good that I created the visualization tools earlier.