2024 Day 2: Red-Nosed Reports
First was easy, the second got a bit tricky with my original
approach to the first. Another solution that would make the
code simpler (but a bit slower), would be to try to remove
each entry in the list and see if there is anyone that results
in a safe report.

Didn't quite think of that since making obviously slow solutions
usually doesn't work later on in AoC. But this solution isn't
too bad and it's pretty quick so I'm happy with it.