2024 Day 3: Mull It Over
Felt like a pretty easy problem. I thought I had both solved in
18 minutes, but then I got the wrong answer for prat two. Took
forever to figure out that I treated the input as completely
self contained lines, which you typically often do. But here
the don't() got reset between lines which caused wrong results.

Very annoying, instead of 3 place on the work leader board I
ended up on the 11th due to this and not being able to spend
more time on the problem until late in the evening.