2024 Day 6: Guard Gallivant
The first part was easy and solved quickly, the second
felt easy as well, but I ran into some bugs that was
hard to track down.

After 5 hours and some help with extra test cases from
Reddit I managed to track down the last mistake and
solve the problem. My last problem was that I added
the obstacle outside of the lab which resulted in
a off by one error in the result.