2024 Day 7: Bridge Repair
The first problem was very quick and easy and the
second felt like it should be that as well. However
I had some issue that I just couldn't figure out.

First I thought it was related to overflows, but that
wasn't the case. I also managed to spot some tiny mistake
fairly early, but that still wasn't enough.

After quite some time I figured out that I treated an
equation as correct if I got the final result without
using all operands. The fix was trivial and my original
solution was very close to being correct.