2024 Day 8: Resonant Collinearity
First part was pretty easy, but I thought there might be
antinodes between two antennas as well, so I added code
for that case. Turned out it was needed and in part 2
it became obvious why now, so I could throw away those
two extra lines of code.

Part two was also easy, but still a bit of work to do
for generating a sequence of antinodes instead of
just providing two fixed antinodes. But not particularly
hard.