2024 Day 9: Disk Fragmenter
This was a bit tricky to get right. A lot of chances for
small errors. But also fun to try to optimize afterward.

Initial solution was around a second on part 2, now it's
at 50ms. Part one was around 100ms and now it's 20ms.

Ended up with completely separate solutions for part
1 and 2 to be able to have quick solutions.