2024 Day 11: Plutonian Pebbles
The first part was fairly easy, but it was obvious that my
approach wouldn't work with part two even before seeing
what it was.

I'm not good at all at dynamic programming so part two
required 1-2 hours of thinking to come up with a solution
to attempt. Initially I kept a list of a all stones in memory
but that busted the heap on larger amounts of blinks.

After some more thinking I realized I don't need the full lists
and that the number of stones is enough. With this the solution
was found quickly and without using much memory.

With some refactoring I can see how people who are good at
dynamic programming could solve this in a couple of minutes,
it wasn't much code required for a good solution.

I kind of hope for more dynamic programming puzzles this year
so I can get better at it.