2024 Day 12: Garden Groups
The first part was pretty quick and easy, but the second part was more
tricky.

I first thought about walking on the insdie, but that gets tricky when
there are multiple holes in the same area that shares a corner. Walking
along the edge of the areas on the outside was easier even though it
was a bit tricky to not introduce bugs in the code.