2024: Day 13: Claw Contraption
It was apparent from the start that a brute force solution would not
be possible for part 2. So spent some time thinking and realiezed I
did some equation solving last year (on 24th of December).

Reusing the gaussianElimination method made it really easy to implement
a solution that worked both for part 1 and 2.