2024 Day 14: Restroom Redoubt
First part was simple math, didn't even have to solve any
equations. The second was much more vague and it was hard
to figure out how you can find a christmas tree.

But looking for many robots with 4 other robots around them
turned out to be a good strategy. Very visualization friendly
problem, and I imagine, very hard for an AI to solve.