2024 Day 15: Warehouse Woes
Part one was fairly quick and easy, but part two
was a bit more tricky. Not too difficult but
required a bit of an effort to get it right.

Then refactoring the code into something that
didn't duplicate a lot of code was also another
challenge. Fun day all in all.