2024 Day 18: RAM Run
Both part 1 and 2 were easy today, especially when I
have a ready BFS implementation.

I spent some time on optimizing part two by doing
a binary search for the first time that blocks the
path to the exit. Not completely trivial to write
initially, but in the end it was not much code required
for this day. Both parts also runs in a couple of tens
of milliseconds, so it's pretty fast.