2024 Day 20: Race Condition
Fairly easy day today thankfully. Had a small bug at
first where I didn't include the start position in
the path. But wasn't to hard to find.

Then when I started with part 2 I realized a much
simpler solution that just counts the number of
cheats rather than creating sets of each cheat based
on if it was possible to move through walls.

The key realization for this to be easy was that there
is only one path to the goal, so you don't need to do
any 2d grid search. Just find the path to the goal and
work on a 1d list of positions.