2024 Day 19: Linen Layout
I struggled a lot with this one. First was very
quick and easy. But on the second it never finished
running.

Had to think a lot to find a way to reduce the amount
of calculations. My first approach was to avoid checking
the same things over and over again by combining patterns
that end up with the same thing. So instead of checking
a + b + c, ab + c, a + bc and abc I tried to remember
that any abc string is made up of 4 different possibilities.

But that got tricky when you get a + bcd. After a long time
I thought I had something that worked but each of the 400
lines in the input took between 15 and 60 seconds.

I added some basic memoization to it and could get down total
runtime to 36 minutes, but with the wrong answer. Added more
memoization and got down the runtime to 200ms instead. Another
answer, but still the wrong answer.

Then the next day I came back and started over again with no
special logic, just looking up all combinations, but saving
all results. Very simple solution and ran in 90ms and produced
the right answer.

It shows the power of memoization. I thought some clever logic
would be needed in addition to just memoization. I hope there
will be some more dynamic programming problems the last few days
so I can get some more practice and actually learn to use memoization
from the start.