2024 Day 21: Keypad Conundrum
I had an easy solution ready very quickly, but it turned
out that it failed on the test input. I was looking for
any way of inputing the code and not the shortest.

It was pretty tricky to come up with a way of finding the
shortest variant. Especially since I could predict that
part 2 would be about having many more robots.

But after some time I managed to get a recursive solution
that was working. Part two was not particularly hard, it was
just about adding memoization.

A big lesson here was to use data classes to compare states
instead of normal classes. For normal classes reference
comparision is done by default which means there will be
no matches. Took a while to figure out why the memoization
didn't work at first. But when using a data class everything
worked fine.