2024 Day 22: Monkey Market
The instructions were a bit hard to understand today, but the problem wasn't
that hard to solve.

The trickiest part was to make the implementation not take too much time
on part two. My initial version took a bit over 10 seconds but with some
optimization I could get it down to around 0.5 seconds.

The biggest saving was from summing up bananas while iterating through
the window the first time instead of first generating all windows and
then iterate through it again to count bananas for all different sequences.

Working on bytes instead of Longs and using a data class as a key for
the map also helped a lot. I initially used a string, but it turned out
it was much faster working with a data class instead.

A bit hard to know exactly what kind of changes makes things faster,
so there were a lot of trial and error involved.