2024 Day 17: Chronospatial Computer
Part 1 was quick and easy, but part two was really hard.

It was fairly easy to reverse engineer the program and
understand what it was doing. But it was tricky to figure out
how to generate the desired input after that. After quite some
thinking I came up with a rough idea that required several attempts
before I managed to implement it.

Interestingly enough the implementation was very short (~10 lines)
once I managed to get it working.

Fun problem.