2024 Day 23: LAN Party
First part wasn't too hard, but required a bit of thinking to
get right. The second part was a bit harder, but I noticed that
on the example input all nodes had 4 neighbours and in the real input
all nodes had 13 neighbours. In the example input there were one group
of nodes with four completely connected neighbours.

So I just assumed that the real input would behave similarly and have
the largest group of size 13. It was the case so I could just make a
simple implementation that found groups of connected neighbours of
size 13 instead of writing some fancy clique generation algorithm.