2025 Day 1: Secret Entrance
First part was pretty easy, but the second was hard to get right.
My initial solution was to just brute force it by turning one step
at a time. But then I rewrote it into the current solution where
it first turns to 0/99 in one go, does a one position turn to wrap
around and then turn the remaining distance (or to 0/99 again).

This made it much easier to understand what's happening and to make
sure the code is actually working.