2025 Day 7: Laboratories
Part one was pretty easy to solve with a BFS solution. Part two
did not work at all with that due to the high number of branches.

Luckily I remembered about using memorization from previous years
and doing a DFS with memorization is easy and sovles it easily. I
ended up also using DFS for part one since the code became more
compact and similar to part 2 with that approach.