2025 Day 9: Movie Theater
First part was a bit more difficult than previous days, but nothing too bad.

Second part was much rougher. I had an idea early on from vague memories
of earlier years (last year maybe?) where I checked if I was inside or
outside of a shape by counting how many times I crossed the border.

I thought that doing the same could be pretty efficient, but I didn't
remember exactly how I implemented it so it was pretty tricky. It also
turned out it's not at all as fast as I thought. It takes one minute
to find the answer. There must be a much better way of doing this
that I'm not seeing.

I definitely will have to get back to this and clean up the code
and try to find a better solution.

But pretty fun day anyway.