2025 Day 10: Factory
Part one was pretty easy with a straight forward BFS but part two
was a nighmare. I tried a lot of different search approaches first
but none worked.

So I had to finally resort to using math which took a lot of time
to understand what I could do. But with some youtube videos and a
lot of discussions with Gemini I finally figured out how things
works and what I needed to implement. It wasn't that easy to get
the implementation right still, but it's a pretty good solution
in the end (I don't like that I have to estimate an upper bound though).